# Conflicts:
#   src/routes/groupchat.jsx

+++ BE/src/main/java/com/ssafy/withview/controller/LoginController.java
@@ -3,6 +3,7 @@ package com.ssafy.withview.controller;
 import java.util.HashMap;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.http.HttpStatus;
@@ -17,6 +18,7 @@ import com.ssafy.withview.constant.Role;
 import com.ssafy.withview.dto.AccessTokenDto;
 import com.ssafy.withview.dto.LoginDto;
 import com.ssafy.withview.dto.UserDto;
+import com.ssafy.withview.exception.BadRequestException;
 import com.ssafy.withview.service.JwtService;
 import com.ssafy.withview.service.LoginService;
 
@@ -48,6 +50,8 @@ public class LoginController {
 			// 로그인 유저 정보
 			UserDto userDto = loginService.getUserInfo(loginDto.getId());
 			log.debug("UserInfo pk: {}", userDto.getSeq());
+			loginService.checkDuplicateLogin(userDto.getSeq()); // 중복 로그인 방지
+			loginService.hasBeenWithdrawn(userDto.getSeq()); // 탈퇴한 회원 로그인 방지
 			Authentication authentication = jwtService.createAuthentication(userDto.getSeq(), loginDto.getPassword());
 			if (authentication != null) {
 				// AccessToken 발급
@@ -68,7 +72,7 @@ public class LoginController {
 			resultMap.put("success", false);
 			resultMap.put("message", e.getMessage());
 			status = HttpStatus.ACCEPTED;
-			log.info("[Error] 일치하는 회원 정보가 없습니다. seq: {}", e.getMessage());
+			log.info("[Error] {}", e.getMessage());
 		} catch (Exception e) {
 			resultMap.put("success", false);
 			resultMap.put("message", e.getMessage());
@@ -85,17 +89,49 @@ public class LoginController {
 	 * @return ResponseEntity (true / false, 상태코드)
 	 */
 	@PostMapping("/logout")
-	public ResponseEntity<Map<String, Object>> logout(@RequestBody UserDto userDto, HttpServletResponse response) {
+	public ResponseEntity<Map<String, Object>> logout(@RequestBody UserDto userDto, HttpServletRequest request,
+		HttpServletResponse response) {
 		log.debug("LoginController - logout: 로그아웃 진행");
 		Map<String, Object> resultMap = new HashMap<>();
 		HttpStatus status;
 		try {
 			log.debug("seq: {}", userDto.getSeq());
+			String accessToken = jwtService.resolveAccessToken(request.getHeader("Authorization"));
+			LoginDto loginDto = jwtService.getLoginInfo(accessToken);
+			if (loginDto.getUserSeq() != userDto.getSeq()) {
+				throw new BadRequestException("BAD_REQUEST");
+			}
 			// Redis 에서 RefreshToken 삭제
 			jwtService.removeRefreshToken(userDto.getSeq());
 			// Cookie 삭제
 			jwtService.removeCookie("RefreshToken");
 			response.setHeader("Set-Cookie", jwtService.removeCookie("RefreshToken").toString());
+			resultMap.put("success", true);
+			status = HttpStatus.OK;
+			log.info("로그아웃 성공. seq: {}", userDto.getSeq());
+		} catch (Exception e) {
+			resultMap.put("success", false);
+			resultMap.put("message", e.getMessage());
+			status = HttpStatus.INTERNAL_SERVER_ERROR;
+			log.error("[Error] 로그아웃 실패: {}", e.getMessage());
+		}
+		return new ResponseEntity<>(resultMap, status);
+	}
+
+	/**
+	 * 로그아웃 (기존 pc 로그아웃)
+	 * @param loginDto (로그인 id)
+	 * @return ResponseEntity (true / false, 상태코드)
+	 */
+	@PostMapping("/logout2")
+	public ResponseEntity<Map<String, Object>> logout2(@RequestBody LoginDto loginDto) {
+		log.debug("LoginController - logout2: 기존 pc 로그아웃 진행");
+		Map<String, Object> resultMap = new HashMap<>();
+		HttpStatus status;
+		try {
+			log.debug("id: {}", loginDto.getId());
+			UserDto userDto = loginService.getUserInfo(loginDto.getId());
+			jwtService.removeRefreshToken(userDto.getSeq()); // Redis 에서 RefreshToken 삭제
 			resultMap.put("success", true);
 			status = HttpStatus.OK;
 			log.info("로그아웃 성공. seq: {}", userDto.getSeq());
